+++ packages/__vital__/src/styleguide/Page/StyleGuideTemplate/index.tsx
@@ -20,6 +20,7 @@ import { Header } from './Header';
 import { FlowContainer } from './FlowContainer';
 import { Images } from './Images';
 import { Footer } from './Footer';
+import { Table } from './Table';
 import { List } from './List';
 import { _default } from './_default';
 
@@ -34,5 +35,6 @@ export const __vital__StyleGuideTemplate = {
   Images,
   FlowContainer,
   Footer,
+  Table,
   List,
 };