+++ packages/__vital__/src/styleguide/Page/index.tsx
@@ -28,6 +28,7 @@ const {
   FlowContainer,
   Images,
   Footer,
+  Table,
   List,
   _default
 } = __vital__StyleGuideTemplate;
@@ -45,6 +46,7 @@ const Default = asFluidToken({
     FlowContainer,
     Images,
     Footer,
+    Table,
     List
   },
 });