+++ packages/vital-list/src/components/TitleInfographic/index.static.ts
@@ -13,6 +13,6 @@
  */
 
 export {
-  staticTokenCollection as vitalTitalInfographic,
+  staticTokenCollection as vitalTitleInfographic,
   StaticBlock as TitleInfographicClean,
 } from '@bodiless/hydration';